Refactor React components for better maintainability
