+++ src/components/Search/SearchPage.tsx
@@ -1,11 +1,11 @@
-import { FC, useState, useEffect, useMemo } from 'react';
+import { FC, useState, useMemo } from 'react';
 import { observer } from 'mobx-react-lite';
-import { Search, X, Tag as TagIcon, Filter, ArrowLeft, ChevronDown, ChevronUp, MoreHorizontal } from 'lucide-react';
+import { ArrowLeft } from 'lucide-react';
 import { useStore } from '../../stores/StoreProvider';
-import { Note, TagNode } from '../../types';
-import { generateUniqueId } from '../../utils';
-import { TagTreeItem } from '../Sidebar/TagTreeItem';
-import { themes } from '../../constants';
+import { Note } from '../../types';
+import { SearchInput } from './SearchInput';
+import { SearchFilters } from './SearchFilters';
+import { SearchResults } from './SearchResults';
 import './SearchPage.css';
 
 interface SearchPageProps {
@@ -19,13 +19,10 @@ interface TagFilter {
 }
 
 export const SearchPage: FC<SearchPageProps> = observer(({ onClose, onNoteSelect }) => {
-  const { notesStore, settingsStore, tagsStore } = useStore();
+  const { notesStore, settingsStore } = useStore();
   const [searchQuery, setSearchQuery] = useState('');
   const [tagFilters, setTagFilters] = useState<TagFilter[]>([]);
-  const [isTagSelectorOpen, setIsTagSelectorOpen] = useState(false);
   const [selectedTagOperator, setSelectedTagOperator] = useState<'AND' | 'OR' | 'NOT'>('AND');
-  const [isFiltersCollapsed, setIsFiltersCollapsed] = useState(false);
-  const [loadingNotes, setLoadingNotes] = useState<Set<string>>(new Set());
 
   // Fuzzy search function
   const fuzzyMatch = (text: string, query: string): boolean => {
@@ -34,7 +31,6 @@ export const SearchPage: FC<SearchPageProps> = observer(({ onClose, onNoteSelect
     const normalizedText = text.toLowerCase();
     const normalizedQuery = query.toLowerCase();
     
-    // Simple fuzzy matching - check if all characters in query appear in order
     let queryIndex = 0;
     for (let i = 0; i < normalizedText.length && queryIndex < normalizedQuery.length; i++) {
       if (normalizedText[i] === normalizedQuery[queryIndex]) {
@@ -48,11 +44,9 @@ export const SearchPage: FC<SearchPageProps> = observer(({ onClose, onNoteSelect
   const getMatchingParagraphs = (content: string, tagFilters: TagFilter[]): string[] => {
     if (tagFilters.length === 0) return [];
     
-    // Create a temporary div to parse HTML content
     const tempDiv = document.createElement('div');
     tempDiv.innerHTML = content;
     
-    // Get all paragraph-like elements
     const elements = tempDiv.querySelectorAll('p, h1, h2, h3, h4, h5, h6, li');
     const matchingParagraphs: string[] = [];
     
@@ -60,10 +54,8 @@ export const SearchPage: FC<SearchPageProps> = observer(({ onClose, onNoteSelect
       const dataTags = element.getAttribute('data-tags') || '';
       if (!dataTags) return;
       
-      // Parse paragraph tags
       const paragraphTags = dataTags.split(',').map(tag => tag.trim()).filter(tag => tag);
       
-      // Check if paragraph tags match any of the tag filters
       const matches = tagFilters.some(filter => {
         return paragraphTags.some(tag => tag.includes(filter.path));
       });
@@ -80,11 +72,9 @@ export const SearchPage: FC<SearchPageProps> = observer(({ onClose, onNoteSelect
   const getTextMatchingParagraphs = (content: string, query: string): string[] => {
     if (!query.trim()) return [];
     
-    // Create a temporary div to parse HTML content
     const tempDiv = document.createElement('div');
     tempDiv.innerHTML = content;
     
-    // Get all paragraph-like elements
     const elements = tempDiv.querySelectorAll('p, h1, h2, h3, h4, h5, h6, li');
     const matchingParagraphs: string[] = [];
     
@@ -99,53 +89,6 @@ export const SearchPage: FC<SearchPageProps> = observer(({ onClose, onNoteSelect
     return matchingParagraphs;
   };
 
-  // Check if note title or full content matches query
-  const noteFullyMatches = (note: Note, query: string): boolean => {
-    if (!query.trim()) return true;
-    
-    const titleMatches = fuzzyMatch(note.title, query);
-    if (titleMatches) return true;
-    
-    // Check if a significant portion of the content matches
-    const tempDiv = document.createElement('div');
-    tempDiv.innerHTML = note.content;
-    const fullText = tempDiv.textContent || '';
-    
-    // If the full text matches, consider it a full match
-    return fuzzyMatch(fullText, query);
-  };
-
-  // Get content to display for a note (either full content or matching paragraphs)
-  const getNoteDisplayContent = (note: Note, query: string): { content: string; isPartial: boolean } => {
-    const filteredContent = getFilteredContent(note);
-    
-    // Check if we have tag filters that match paragraphs
-    const matchingTagParagraphs = getMatchingParagraphs(filteredContent, tagFilters);
-    
-    if (!query.trim() && tagFilters.length === 0) {
-      return { content: filteredContent, isPartial: false };
-    }
-    
-    if (query.trim() && noteFullyMatches(note, query) && tagFilters.length === 0) {
-      return { content: filteredContent, isPartial: false };
-    }
-    
-    // Combine text and tag matching paragraphs
-    const textMatchingParagraphs = query.trim() ? getTextMatchingParagraphs(filteredContent, query) : [];
-    
-    // Merge both types of matching paragraphs (remove duplicates)
-    const allMatchingParagraphs = [...new Set([...textMatchingParagraphs, ...matchingTagParagraphs])];
-    
-    if (allMatchingParagraphs.length > 0) {
-      // Join matching paragraphs with separators
-      const partialContent = allMatchingParagraphs
-        .join('<br/><br/>');
-      return { content: partialContent, isPartial: true };
-    }
-    
-    return { content: filteredContent, isPartial: false };
-  };
-
   // Filter notes based on search query and tag filters
   const filteredNotes = useMemo(() => {
     let notes = notesStore.getVisibleNotes(settingsStore.isCensorshipEnabled());
@@ -153,12 +96,10 @@ export const SearchPage: FC<SearchPageProps> = observer(({ onClose, onNoteSelect
     // Apply text search
     if (searchQuery.trim()) {
       notes = notes.filter(note => {
-        // Check title match
         if (fuzzyMatch(note.title, searchQuery.trim())) {
           return true;
         }
         
-        // Check if any paragraphs match
         const matchingParagraphs = getTextMatchingParagraphs(note.content, searchQuery.trim());
         return matchingParagraphs.length > 0;
       });
@@ -167,31 +108,25 @@ export const SearchPage: FC<SearchPageProps> = observer(({ onClose, onNoteSelect
     // Apply tag filters
     if (tagFilters.length > 0) {
       notes = notes.filter(note => {
-        // First check if note has matching paragraph tags
         const matchingParagraphs = getMatchingParagraphs(note.content, tagFilters);
         if (matchingParagraphs.length > 0) {
           return true;
         }
         
-        // Then check note-level tags
         const noteTags = note.tags.map(tag => tag.path);
         
-        // Group filters by operator
         const andFilters = tagFilters.filter(f => f.operator === 'AND');
         const orFilters = tagFilters.filter(f => f.operator === 'OR');
         const notFilters = tagFilters.filter(f => f.operator === 'NOT');
 
-        // AND filters - all must match
         const andMatch = andFilters.length === 0 || andFilters.every(filter =>
           noteTags.some(tag => tag.includes(filter.path))
         );
 
-        // OR filters - at least one must match
         const orMatch = orFilters.length === 0 || orFilters.some(filter =>
           noteTags.some(tag => tag.includes(filter.path))
         );
 
-        // NOT filters - none must match
         const notMatch = notFilters.every(filter =>
           !noteTags.some(tag => tag.includes(filter.path))
         );
@@ -200,7 +135,7 @@ export const SearchPage: FC<SearchPageProps> = observer(({ onClose, onNoteSelect
       });
     }
 
-    // Sort by relevance (title matches first, then by creation date)
+    // Sort by relevance
     return notes.sort((a, b) => {
       if (searchQuery.trim()) {
         const aTitle = a.title.toLowerCase().includes(searchQuery.toLowerCase());
@@ -214,25 +149,10 @@ export const SearchPage: FC<SearchPageProps> = observer(({ onClose, onNoteSelect
     });
   }, [searchQuery, tagFilters, notesStore.notes, settingsStore.isCensorshipEnabled()]);
 
-  // Filter content based on censorship mode
-  const getFilteredContent = (note: Note): string => {
-    if (!settingsStore.isCensorshipEnabled()) {
-      return note.content;
-    }
-    
-    // Remove censored content when censorship is enabled
-    const tempDiv = document.createElement('div');
-    tempDiv.innerHTML = note.content;
-    const censoredElements = tempDiv.querySelectorAll('span[data-censored]');
-    censoredElements.forEach(el => el.remove());
-    return tempDiv.innerHTML;
-  };
-
   const addTagFilter = (tagPath: string) => {
     if (!tagFilters.some(f => f.path === tagPath)) {
       setTagFilters([...tagFilters, { path: tagPath, operator: selectedTagOperator }]);
     }
-    setIsTagSelectorOpen(false);
   };
 
   const removeTagFilter = (tagPath: string) => {
@@ -250,130 +170,6 @@ export const SearchPage: FC<SearchPageProps> = observer(({ onClose, onNoteSelect
     setTagFilters([]);
   };
 
-  const handleNoteClick = async (note: Note) => {
-    // Load note content if not already loaded
-    if (!note.content) {
-      setLoadingNotes(prev => new Set(prev).add(note.id));
-      await notesStore.loadNoteContent(note);
-      setLoadingNotes(prev => {
-        const newSet = new Set(prev);
-        newSet.delete(note.id);
-        return newSet;
-      });
-    }
-    onNoteSelect(note);
-  };
-
-  const getOperatorColor = (operator: 'AND' | 'OR' | 'NOT') => {
-    switch (operator) {
-      case 'AND': return '#22c55e';
-      case 'OR': return '#3b82f6';
-      case 'NOT': return '#ef4444';
-      default: return '#6b7280';
-    }
-  };
-
-  // Load content for all filtered notes
-  useEffect(() => {
-    const loadAllContent = async () => {
-      const notesToLoad = filteredNotes.filter(note => !note.content);
-      if (notesToLoad.length === 0) return;
-
-      setLoadingNotes(prev => {
-        const newSet = new Set(prev);
-        notesToLoad.forEach(note => newSet.add(note.id));
-        return newSet;
-      });
-
-      // Load content for all notes in parallel
-      await Promise.all(
-        notesToLoad.map(note => notesStore.loadNoteContent(note))
-      );
-
-      setLoadingNotes(prev => {
-        const newSet = new Set(prev);
-        notesToLoad.forEach(note => newSet.delete(note.id));
-        return newSet;
-      });
-    };
-
-    loadAllContent();
-  }, [filteredNotes, notesStore]);
-
-  const renderNoteContent = (note: Note) => {
-    const notebook = notesStore.notebooks.find(nb => nb.id === note.notebookId);
-    const isLoading = loadingNotes.has(note.id);
-    const { content: displayContent, isPartial } = getNoteDisplayContent(note, searchQuery);
-    
-    
-    // Apply note-specific theme styles
-    const noteTheme = note.theme ? themes[note.theme]?.settings : settingsStore.settings;
-    const noteStyles = noteTheme ? {
-      fontFamily: noteTheme.editorFontFamily,
-      fontSize: noteTheme.editorFontSize,
-      lineHeight: noteTheme.editorLineHeight,
-    } : {};
-
-    return (
-      <div key={note.id} className="search-note-item">        
-        <div 
-          className="search-note-header" 
-          onClick={() => handleNoteClick(note)}
-          data-note-id={note.id}
-        >
-          <h2 style={{ 
-            fontFamily: noteTheme?.titleFontFamily || settingsStore.settings.titleFontFamily,
-            fontSize: noteTheme?.titleFontSize || settingsStore.settings.titleFontSize
-          }}>
-            {note.title}
-          </h2>
-        </div>
-        
-        <div 
-          className="search-note-content"
-          data-partial={isPartial}
-          style={noteStyles}
-        >
-          {isLoading ? (
-            <div className="note-loading">
-              <div className="loading-spinner-small"></div>
-              <span>Loading content...</span>
-            </div>
-          ) : displayContent ? (
-            <div 
-              dangerouslySetInnerHTML={{ 
-                __html: displayContent
-              }}
-              onClick={(e) => {
-                // Prevent event bubbling to note header
-                e.stopPropagation();
-              }}
-            />
-          ) : (
-            <div className="no-content">
-              <em>No content</em>
-            </div>
-          )}
-        </div>
-      </div>
-    );
-  };
-
-  const renderTagTree = (nodes: TagNode[], level = 0) => {
-    return nodes.map(node => (
-      <div key={node.id} style={{ paddingLeft: `${level}rem` }}>
-        <div 
-          className="tag-selector-item"
-          onClick={() => addTagFilter(node.name)}
-        >
-          <TagIcon className="h-4 w-4" />
-          <span>{node.name}</span>
-        </div>
-        {node.children.length > 0 && renderTagTree(node.children, level + 1)}
-      </div>
-    ));
-  };
-
   return (
     <div className="search-page">
       <div className="search-header">
@@ -385,173 +181,28 @@ export const SearchPage: FC<SearchPageProps> = observer(({ onClose, onNoteSelect
       </div>
 
       <div className="search-content">
-        <div className={`search-panel ${isFiltersCollapsed ? 'collapsed' : ''}`}>
-          <button 
-            className="filters-toggle"
-            onClick={() => setIsFiltersCollapsed(!isFiltersCollapsed)}
-          >
-            <Filter className="h-4 w-4" />
-            <span>Filters</span>
-            {isFiltersCollapsed ? (
-              <ChevronDown className="h-4 w-4" />
-            ) : (
-              <ChevronUp className="h-4 w-4" />
-            )}
-          </button>
-          
-          <div className="search-input-section">
-            <div className="search-input-wrapper">
-              <Search className="h-5 w-5 search-icon" />
-              <input
-                type="text"
-                value={searchQuery}
-                onChange={(e) => setSearchQuery(e.target.value)}
-                placeholder="Search notes... (fuzzy search supported)"
-                className="search-input"
-                autoFocus
-              />
-              {searchQuery && (
-                <button
-                  onClick={() => setSearchQuery('')}
-                  className="search-clear-button"
-                >
-                  <X className="h-4 w-4" />
-                </button>
-              )}
-            </div>
-          </div>
+        <SearchInput
+          searchQuery={searchQuery}
+          onSearchChange={setSearchQuery}
+        />
+        
+        <SearchFilters
+          tagFilters={tagFilters}
+          onAddTagFilter={addTagFilter}
+          onRemoveTagFilter={removeTagFilter}
+          onUpdateTagOperator={updateTagOperator}
+          onClearAllFilters={clearAllFilters}
+          selectedTagOperator={selectedTagOperator}
+          onSetSelectedTagOperator={setSelectedTagOperator}
+          searchQuery={searchQuery}
+        />
 
-          {!isFiltersCollapsed && (
-            <div className="tag-filters-section">
-            <div className="tag-filters-header">
-              <h3>Tag Filters</h3>
-              <div className="tag-operator-selector">
-                <label>New filter type:</label>
-                <select
-                  value={selectedTagOperator}
-                  onChange={(e) => setSelectedTagOperator(e.target.value as 'AND' | 'OR' | 'NOT')}
-                  className="operator-select"
-                >
-                  <option value="AND">AND (must have)</option>
-                  <option value="OR">OR (can have)</option>
-                  <option value="NOT">NOT (must not have)</option>
-                </select>
-              </div>
-            </div>
-
-            <div className="active-filters">
-              {tagFilters.map((filter, index) => (
-                <div key={index} className="tag-filter-chip">
-                  <span 
-                    className="filter-operator"
-                    style={{ backgroundColor: getOperatorColor(filter.operator) }}
-                  >
-                    {filter.operator}
-                  </span>
-                  <span className="filter-tag">{filter.path}</span>
-                  <select
-                    value={filter.operator}
-                    onChange={(e) => updateTagOperator(filter.path, e.target.value as 'AND' | 'OR' | 'NOT')}
-                    className="filter-operator-select"
-                  >
-                    <option value="AND">AND</option>
-                    <option value="OR">OR</option>
-                    <option value="NOT">NOT</option>
-                  </select>
-                  <button
-                    onClick={() => removeTagFilter(filter.path)}
-                    className="filter-remove-button"
-                  >
-                    <X className="h-3 w-3" />
-                  </button>
-                </div>
-              ))}
-            </div>
-
-            <div className="tag-selector">
-              <button
-                onClick={() => setIsTagSelectorOpen(!isTagSelectorOpen)}
-                className="add-tag-filter-button"
-              >
-                <TagIcon className="h-4 w-4" />
-                Add Tag Filter
-              </button>
-
-              {isTagSelectorOpen && (
-                <div className="tag-selector-dropdown">
-                  <div className="tag-selector-content">
-                    {tagsStore.tagTree.length > 0 ? (
-                      renderTagTree(tagsStore.tagTree)
-                    ) : (
-                      <p className="no-tags-message">No tags available</p>
-                    )}
-                  </div>
-                </div>
-              )}
-            </div>
-
-            {(searchQuery || tagFilters.length > 0) && (
-              <button onClick={clearAllFilters} className="clear-filters-button">
-                <Filter className="h-4 w-4" />
-                Clear All Filters
-              </button>
-            )}
-          </div>
-          )}
-        </div>
-
-        <div className="search-results">
-          <div className="results-header">
-            <h3>
-              {filteredNotes.length} note{filteredNotes.length !== 1 ? 's' : ''} found
-            </h3>
-            {(searchQuery || tagFilters.length > 0) && (
-              <div className="search-summary">
-                {searchQuery && (
-                  <span className="search-term">
-                    Text: "{searchQuery}"
-                  </span>
-                )}
-                {tagFilters.length > 0 && (
-                  <span className="tag-summary">
-                    Tags: {tagFilters.map(f => `${f.operator} ${f.path}`).join(', ')}
-                  </span>
-                )}
-              </div>
-            )}
-          </div>
-
-          <div className="combined-notes-container">
-            {filteredNotes.length > 0 ? (
-              <div className="search-notes-list">
-                {filteredNotes.map((note, index) => (
-                  <div key={note.id}>
-                    {renderNoteContent(note)}
-                    {index < filteredNotes.length - 1 && (
-                      <hr className="note-separator" />
-                    )}
-                  </div>
-                ))}
-              </div>
-            ) : null}
-
-            {filteredNotes.length === 0 && (searchQuery || tagFilters.length > 0) && (
-              <div className="no-results">
-                <Search className="h-12 w-12 no-results-icon" />
-                <h3>No notes found</h3>
-                <p>Try adjusting your search terms or tag filters</p>
-              </div>
-            )}
-
-            {filteredNotes.length === 0 && !searchQuery && tagFilters.length === 0 && (
-              <div className="no-results">
-                <Search className="h-12 w-12 no-results-icon" />
-                <h3>Start searching</h3>
-                <p>Enter search terms or add tag filters to find your notes</p>
-              </div>
-            )}
-          </div>
-        </div>
+        <SearchResults
+          filteredNotes={filteredNotes}
+          searchQuery={searchQuery}
+          tagFilters={tagFilters}
+          onNoteSelect={onNoteSelect}
+        />
       </div>
     </div>
   );
